Jackson Rest Support + REST-Assured Integration Tests with Jacoco Report Generation (Still 0% coverage)
This commit adds support to Rest Assured + Report Generation. This is still to show
that Jacoco does not work even with the Agent being requested before the integration
tests are executed.

gradle integrationTest jacocoReport

This executes the new REST-assured test cases and generate the jacoco report based on the execution.
It is still generating 0% covered when running from tomcat.